scripts prepare and run added env vars added

+++ models/response.go
@@ -1,7 +1,7 @@
 package models
 
 type Response struct {
-    TotalItems		int		`json:"total_items"`
-    TotalCategories int		`json:"total_categories"`
-    TotalPrice		float64 `json:"total_price"`
+    TotalItems		int	`json:"total_items"`
+    TotalCategories int	`json:"total_categories"`
+    TotalPrice		int `json:"total_price"`
 }
